+++ src/app.ts
@@ -4,21 +4,27 @@ import cors from "cors";
 import cron from "node-cron";
 import path from "path";
 import fs from "fs";
-import { getStatus } from "./get-status";
+import { getStatus } from "./routes/get-status";
 import { getBudgets } from "./routes/get-budgets";
 import { log } from "./utils/debug";
-import getCarouselData from "./get-carousel-data";
-import getPrice from "./get-price";
-import getCirculatingSupply from "./get-circulating-supply";
-import { calculateSecondsUntilNext5MinuteInterval } from "./utils";
-import getTotalSupply from "./get-total-supply";
+import getCarouselData from "./routes/get-carousel-data";
+import getPrice from "./routes/get-price";
+import getCirculatingSupply from "./routes/get-circulating-supply";
+import { calculateSecondsUntilNext5MinuteInterval } from "./utils/utils";
+import getTotalSupply from "./routes/get-total-supply";
 
 const app = express();
 const cache = apicache.middleware;
 const port = process.env.PORT || 8081;
 const CAROUSEL_DATA_PATH = path.join(__dirname, "../carousel-data.json");
-const CIRCULATING_SUPPLY_DATA_PATH = path.join(__dirname, "../circulating-supply-data.json");
-const TOTAL_SUPPLY_DATA_PATH = path.join(__dirname, "../total-supply-data.json");
+const CIRCULATING_SUPPLY_DATA_PATH = path.join(
+  __dirname,
+  "../circulating-supply-data.json"
+);
+const TOTAL_SUPPLY_DATA_PATH = path.join(
+  __dirname,
+  "../total-supply-data.json"
+);
 
 app.use(cors());
 app.use(express.json());
@@ -36,13 +42,22 @@ const scheduleCronJob = async () => {
     const circulatingSupplyData = await getCirculatingSupply();
     const totalSupplyData = await getTotalSupply();
 
-    fs.writeFileSync(CIRCULATING_SUPPLY_DATA_PATH, JSON.stringify(circulatingSupplyData, null, 2));
-    fs.writeFileSync(TOTAL_SUPPLY_DATA_PATH, JSON.stringify(totalSupplyData, null, 2));
+    fs.writeFileSync(
+      CIRCULATING_SUPPLY_DATA_PATH,
+      JSON.stringify(circulatingSupplyData, null, 2)
+    );
+    fs.writeFileSync(
+      TOTAL_SUPPLY_DATA_PATH,
+      JSON.stringify(totalSupplyData, null, 2)
+    );
   };
 
   // Fetch data initially such that we have something to serve. There will at most
   // be a buffer of 5 minutes from this running until the first cron execution.
-  if (!fs.existsSync(CIRCULATING_SUPPLY_DATA_PATH) || !fs.existsSync(TOTAL_SUPPLY_DATA_PATH))
+  if (
+    !fs.existsSync(CIRCULATING_SUPPLY_DATA_PATH) ||
+    !fs.existsSync(TOTAL_SUPPLY_DATA_PATH)
+  )
     await fetchAndWriteSupplyData();
   if (!fs.existsSync(CAROUSEL_DATA_PATH)) await fetchAndWriteCarouselData();
