+++ src/database/team.ts
@@ -1,0 +1,22 @@
+import express from "express";
+import db from "./db";
+
+const app = express();
+
+// create database table team
+app.get("/create/table/team", async (req, res) => {
+  try {
+    const client = await db.connect();
+    await client.query("CREATE SCHEMA IF NOT EXISTS dashbaord");
+    await client.query(`
+      CREATE TABLE IF NOT EXISTS dashbaord.team (
+        id SERIAL PRIMARY KEY,
+        councilmembers VARCHAR(50) NOT NULL,
+        leads VARCHAR(50) NOT NULL,
+        created_at TIMESTAMP NOT NULL DEFAULT NOW()
+      )
+    `);
+  } catch (err) {
+    console.log(err);
+  }
+});
