+++ src/database/telegram.ts
@@ -1,0 +1,22 @@
+import express from "express";
+import db from "./db";
+
+const app = express();
+
+// create database table telegram
+app.get("/create/table/telegram", async (req, res) => {
+  try {
+    const client = await db.connect();
+    await client.query("CREATE SCHEMA IF NOT EXISTS dashbaord");
+    await client.query(`
+      CREATE TABLE IF NOT EXISTS dashbaord.telegram (
+        id SERIAL PRIMARY KEY,
+        usersMembers INTEGER NOT NULL,
+        messageNumbers INTEGER NOT NULL,
+        created_at TIMESTAMP NOT NULL DEFAULT NOW()
+      )
+    `);
+  } catch (err) {
+    console.log(err);
+  }
+});
